backend/requirements/base.txt
# Payless Cars - Base Requirements
# Core Django dependencies shared across all environments

# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# Django Extensions
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
django-environ>=0.11,<1.0

# Database
psycopg2-binary>=2.9,<3.0

# Task Queue
celery>=5.3,<6.0
redis>=5.0,<6.0

# AWS
boto3>=1.34,<2.0
django-storages>=1.14,<2.0

# Image Processing
Pillow>=10.0,<11.0

# Production Server
gunicorn>=21.0,<22.0

# Utilities
python-dateutil>=2.8,<3.0
